Add Numba-accelerated 3-turn MDP and fix mdp.py dtype bug
darts/mdp_3turn.py: new module implementing 3-turn-aware value
iteration. Processes states in groups ordered by round_start,
exploiting the monotonic dependency between groups so each group
converges independently. The @njit inner loop matches the per-dart
cost convention of the published single-player MDP.

darts/mdp.py: fix points array dtype (int64 → int32) so the Numba
typed dict in _compute_actions can accept point values as values.

notebooks/MDP 3-turn 7Feb.ipynb: updated to use the new module.
Profiling showed quadratic scaling: game_start=501 runs in ~8 min
(20s→1.6s, 50→8s, 100→25s, 200→83s, 501→485s measured).

+++ darts/mdp.py
@@ -275,7 +275,8 @@ class SinglePlayerContinuousMDP:
                 )
                 if np.linalg.norm(np.array([i, j]) - self.centre)
                 < radius_pixels + self.margin
-            ]
+            ],
+            dtype=np.int32,
         )
 
     def compute_transition_probs_from_point(self, point):
